Se agregaron validaciones

+++ src/main/java/com/proyectTest/proyectTest/controller/AppointmentController.java
@@ -31,37 +31,43 @@ public class AppointmentController {
     }
 
     @GetMapping("/dentist/{id}")
-    public ResponseEntity<List<AppointmentDto>> getAllByDentistId(@PathVariable Integer id){
+    public ResponseEntity<List<AppointmentDto>> getAllByDentistId(@PathVariable Long id){
         return ResponseEntity.ok(appointmentService.getAllByDentistId(id));
     }
 
     @GetMapping("/dentist/{id}/date/{date}")
-    public ResponseEntity<List<AppointmentDto>> getAllByDentistIdAndDate(@PathVariable int id, @PathVariable String date){
+    public ResponseEntity<List<AppointmentDto>> getAllByDentistIdAndDate(@PathVariable Long id, @PathVariable String date){
         return ResponseEntity.ok(appointmentService.getAllByDentistIdAndDate(id, date));
     }
 
     @GetMapping("/patient/{id}")
-    public ResponseEntity<List<AppointmentDto>> getAllByPatientId(@PathVariable Integer id){
+    public ResponseEntity<List<AppointmentDto>> getAllByPatientId(@PathVariable Long id){
         return ResponseEntity.ok(appointmentService.getAllByPatientId(id));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<AppointmentDto> getById(@PathVariable Integer id){
+    public ResponseEntity<AppointmentDto> getById(@PathVariable Long id) {
         AppointmentDto appointmentDto = appointmentService.getById(id).orElse(null);
         return ResponseEntity.ok(appointmentDto);
     }
 
-
-
     @PutMapping()
     public ResponseEntity<AppointmentDto> update(@RequestBody Appointment appointment) {
-        return ResponseEntity.ok(appointmentService.update(appointment));
+        if(appointment.getId() != null && appointmentService.getById(appointment.getId()) != null) {
+            return ResponseEntity.ok(appointmentService.update(appointment));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> delete(@PathVariable Integer id) {
-        appointmentService.delete(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Patient deleted");
+    public ResponseEntity<String> delete(@PathVariable Long id) {
+        if(appointmentService.getById(id) != null) {
+            appointmentService.delete(id);
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Turno eliminado");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
     }
 
 }
